increased compatibility with python3 or python

+++ run.py
@@ -113,7 +113,8 @@ for file_ in files:
 
 from scripts.createReport import endReport
 endReport(f"{current_date}_{current_time}")
-os.remove("./merged_res.csv")
+if os.path.exists("./merged_res.csv"):
+    os.remove("./merged_res.csv")
 print(f"{bcolors.OKGREEN} All jobs finished! {bcolors.ENDC}")
